Open the device at the rate it will really run, and resample with soxr
Closes the bug the previous commit only diagnosed. miniaudio accepts a rate it
has no intention of running, keeps it in device.sampleRate, runs the hardware at
internalSampleRate, and puts a linear resampler between the two. Nothing
downstream could see it, because negotiatedFormat() reported the rate that had
been asked for -- so AudioConverter compared two equal numbers, correctly did
nothing, and soxr sat unused above a linear resampler. On a 44.1 kHz device in
shared mode, which is the shipping default, that was every 48 kHz and higher
track.

IAudioOutput::effectiveSampleRate(wanted, deviceId, exclusive) is the question
that was missing, and it is deliberately not supportsSampleRate(): that one asks
whether a rate can be requested, this asks what the hardware will be doing
afterwards. A backend is entitled to say yes to the first and mean nothing by
it.

play() asks it after the existing refused-rate fallback and before it builds
anything that has to agree about the rate -- the converter, the chain's
prepare(), and both rings, all filled before the device opens. Reading the truth
back after start() would be too late for every one of them, which is why this is
a new question rather than a read-back. switchOutputDevice() asks the same.

MiniaudioOutput answers in two lines of policy: exclusive returns wanted,
because an exclusive stream owns the device and switches it; shared returns the
rate it is already running, because a shared stream never moves it. Both halves
were measured with tools/ma-rate-probe rather than assumed. A backend that will
not say keeps the caller's rate, since guessing for a device that could not be
described is worse than the resampling this exists to avoid.

Keeping the share mode in the signature is what preserves bit-perfect exclusive
playback, and what keeps DoP reachable at all -- its carrier must be running at
exactly 176,400 or 352,800, and an output that helpfully resampled it would emit
the markers as noise.

latencySeconds() was the same confusion and is fixed alongside: it divided
internalPeriodSizeInFrames, a hardware-rate quantity, by the requested rate. It
now takes both numbers from the same place.

Still not covered, and recorded as such: exclusive asked for and refused.
start() falls back to shared on its own, by which point the converter is built.
Closing that means splitting start() into an open that negotiates and a start
that begins pulling, which is not worth the churn on its own.

Three tests against a double for what the engine does with the answer, and a
hidden [.ratedevice] case that asks the real backend the same questions --
because a double can only prove the engine's half, never that the answer is true
of the machine. Also fixes a -Wdouble-promotion that had crept back in; the tree
being warning-free is only worth anything while it is actually true.

+++ app/src/MainFrame.cpp
@@ -569,7 +569,11 @@ void MainFrame::wireUp() {
         // Through the slider rather than straight to the engine, so the panel and
         // the window cannot end up showing different volumes.
         volume_->SetValue(static_cast<int>(std::lround(gain * 100.0F)));
-        playback_->setVolume(gain);
+        // Widened explicitly. MPRIS carries a float and setVolume() takes a
+        // double, so the conversion happens either way; saying so is what keeps
+        // -Wdouble-promotion quiet, and the tree warning-free is a property that
+        // is only worth anything while it is actually true.
+        playback_->setVolume(static_cast<double>(gain));
     });
     observe(media_->openUrlRequested, [this](const Url& url) { openUrls({url}); });
